Add basic Object Detection algorithm
Block scanner currently scans angles [0,90] for a max of two blocks.
If more than two blocks are detected, we stop searching. We first move
to the first block to validate it with the color sensor. If that fails,
we reverse until we reach the origin again. We repeat the process for the
second block if there is one.

Errors:
- The current algorithm does not scan the entire 90 degrees before validating
  a block. I used to have an ArrayOutOfBoundsException in place of this. I
  believe this has to do with the way I am iterating through my scanner loop.

TODO:
- If only one block is detected (and it isn't the styrofoam), move to the
  destination point and repeat the scanning process for angles [180,270].
- Obstacle avoidance (implement wall follower from lab 3).